requirements.txt
pyinstaller
labjack-ljm
ttkbootstrap